dev-requirements.txt
flake8
pytest
pytest-socket
pyfakefs
aioresponses
pytest-coverage
aiohttp-devtools
